remove intermediate conversations

+++ vllm/model_executor/models/mixtral.py
@@ -110,7 +110,7 @@ class MixtralMoE(nn.Module):
         self.s2 = nn.Parameter(
             torch.empty(self.intermediate_size,
                         device="cuda",
-                        dtype=torch.float16))
+                        dtype=torch.float8_e4m3fn))
 
         set_weight_attrs(self.ws, {
             "weight_loader": self.weight_loader,
